--- src/components/deck.js
@@ -1,12 +1,0 @@
-import Card from "./card";
-
-function Deck({ deckList }) {
-  const deck = deckList.map((n, index) => <Card key={index} card={n}/>)
-  return (
-    <div className="Deck">
-      {deck}
-    </div>
-  )
-}
-
-export default Deck
